\name{easyrsqlite-package}
\alias{easyrsqlite-package}
\alias{easyrsqlite}
\docType{package}
\title{Simple database interface for SQLite.}
\description{Simple database interface for SQLite. easyrsqlite is build on top of RSQLite. It provides functions for reading, writing, listing and subsetting tables in SQLite. easyrsqlite allow to retrieve, write and subset database's tables in a similar way as data.frame's, using single square brackets db[,,,]. easyrsqlite internalise many database operations (otherwise tedious) such as rows and columns subsetting (no SQL needed) and connection and disconnection to / from databases.}
\details{
\tabular{ll}{Package: \tab easyrsqlite\cr
Version: \tab 0.7.4\cr
Date: \tab 2014-09-30\cr
Title: \tab Simple database interface for SQLite.\cr
Author: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Maintainer: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Depends: \tab R (>= 3.1.1), RSQLite, easydb\cr
Suggests: \tab \cr
License: \tab AGPL-3\cr
URL: \tab http://easydb.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <Julien.Moeys@mark.slu.se>}

\keyword{ package }


